processIcon helpers
